requirement.txt
streamlit
PyPDF2
sentence-transformers
transformers
faiss-cpu
numpy
